pyproject.toml
[build-system]
requires = ['setuptools', 'cmake>=3.22', 'numpy >= 2']
build-backend = 'setuptools.build_meta'

[project]
name = 'teds'
version = '0.0.1'
requires-python = ">= 3.11"
dependencies = [
    'netCDF4',
    'h5py',
    'scipy',
    'numpy',
    'PyYAML',
    'astropy',
    'sgp4',
    'geopandas',
    'pystac_client == 0.6.1',
    'tqdm',
    'rioxarray',
    'numba',
    'shapely',
    'threadpoolctl',
    'tables',
    'pyproj',
    'Glymur == 0.13.2',
    'pyquaternion == 0.9.9',
    'piccolo_theme == 0.22',
    'flake8 == 7.1.1',
    'mypy == 1.11.1',
    'types-pyyaml',
    'types-tqdm == 4.66',
    'types-requests',
    'sphinx == 7.3',
    'piccolo_theme == 0.22.0',
    'pytest == 8.3',
    'coverage == 7.6',
    'pandas-stubs',
]

[tool.setuptools]
packages = ['teds']

[tool.coverage.run]
omit = ['teds/lib/hapi.py']

[tool.mypy]
disallow_untyped_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = [
    'netCDF4',
    'shapely',
    'geopandas',
    'scipy.*',
    'numba.*',
    'cartopy.*',
    'matplotlib.*',
    'mpl_toolkits.*',
    'astropy.*',
    'sgp4.*',
    'pyquaternion.*',
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    'teds.lib.hapi',
    'teds.lib.libRT',
    'teds.lib.data_netcdf.*',
    'teds.lib.libINV',
    'teds.lib.libAMF',
    'teds.lib.libDOAS',
    'teds.lib.libNumTools',
    'teds.lib.libRT_no2',
    'teds.lib.libWrite',
    'teds.lib.lib_utils',
    'teds.lib.remotap_preproc.*',
    'teds.sgm.sgm_no2',
    'teds.sgm.sgm_geoscene_to_remotap',
    'teds.im.create_im_configuration_file_nitro',
    'teds.ckd.ckd_generation.ckd_generator_nitro',
    'teds.ckd.ckd_generation.generator_class',
    'teds.ckd.ckd_generation.nitro_ckds.*',
    'teds.ckd.create_binning_tables',
    'teds.l1al1b.create_l1a1b_configuration_file_nitro',
    'teds.l1l2.l1bl2_no2',
    'teds.l1l2.l1bl2',
    'teds.l1l2.l1l2_pyTorch',
    'teds.l2l4.ModuleCFM',
    'teds.l2l4.ModuleDataContainer',
    'teds.l2l4.ModuleLSQ',
    'teds.l2l4.ModuleLevel2',
    'teds.l2l4.ModulePlume',
    'teds.l2l4.ModulePreProcessing',
    'teds.l2l4.ModuleReadMicroHH',
    'teds.l2l4.ModuleRemoveBackground',
    'teds.l2l4.l2l4',
    'teds.siml1b.radiance_offset',
    'teds.run_E2E',
]
ignore_errors = true
